Creating jaxrs-client with scala support
